+++ packages/hedron-workbench/src/hedron_workbench/__init__.py
@@ -41,7 +41,7 @@ from hedron_posit import (
 )
 from hedron_workbench.app import HedronWorkbench
 
-__version__ = "0.60.1"
+__version__ = "0.60.2"
 
 __all__ = [
     "ConnectConfig",